Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
mysql-connector-python==9.0.0
pydantic==2.9.0
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.1